moved code around
